Move parsed tree caching to GeneratorDriver
