+++ src/Compilers/CSharp/Portable/SourceGeneration/CSharpGeneratorDriver.cs
@@ -4,10 +4,8 @@
 
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Runtime.CompilerServices;
 using System.Threading;
 using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.Text;
 
 namespace Microsoft.CodeAnalysis.CSharp
 {
@@ -16,8 +14,6 @@ namespace Microsoft.CodeAnalysis.CSharp
     /// </summary>
     public sealed class CSharpGeneratorDriver : GeneratorDriver
     {
-        private static readonly ConditionalWeakTable<SourceText, SyntaxTree> s_parsedGeneratedSources = new();
-
         /// <summary>
         /// Creates a new instance of <see cref="CSharpGeneratorDriver"/>
         /// </summary>
@@ -55,25 +51,7 @@ namespace Microsoft.CodeAnalysis.CSharp
             => new CSharpGeneratorDriver(parseOptions ?? CSharpParseOptions.Default, generators.ToImmutableArray(), optionsProvider ?? CompilerAnalyzerConfigOptionsProvider.Empty, additionalTexts.AsImmutableOrEmpty());
 
         internal override SyntaxTree ParseGeneratedSourceText(GeneratedSourceText input, string fileName, CancellationToken cancellationToken)
-        {
-            if (s_parsedGeneratedSources.TryGetValue(input.Text, out var existingTree)
-                && Equals(_state.ParseOptions, existingTree.Options)
-                && Equals(fileName, existingTree.FilePath))
-            {
-                return existingTree;
-            }
-
-            var tree = SyntaxFactory.ParseSyntaxTree(input.Text, _state.ParseOptions, fileName, cancellationToken);
-
-#if NETCOREAPP
-            s_parsedGeneratedSources.AddOrUpdate(input.Text, tree);
-#else
-            s_parsedGeneratedSources.Remove(input.Text);
-            s_parsedGeneratedSources.Add(input.Text, tree);
-#endif
-
-            return tree;
-        }
+            => SyntaxFactory.ParseSyntaxTree(input.Text, _state.ParseOptions, fileName, cancellationToken);
 
         internal override GeneratorDriver FromState(GeneratorDriverState state) => new CSharpGeneratorDriver(state);
